refactor: util functions moved to utils.rs
